chore: extract `useUserUpload` hook
